+++ src/main/java/org/gestern/gringotts/Gringotts.java
@@ -1,5 +1,6 @@
 package org.gestern.gringotts;
 
+import static org.gestern.gringotts.Configuration.CONF;
 import static org.gestern.gringotts.dependency.Dependency.DEP;
 
 import java.io.IOException;
@@ -10,14 +11,17 @@ import net.milkbowl.vault.economy.Economy;
 import org.bukkit.Bukkit;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.configuration.file.FileConfiguration;
+import org.bukkit.plugin.PluginManager;
 import org.bukkit.plugin.ServicePriority;
 import org.bukkit.plugin.ServicesManager;
 import org.bukkit.plugin.java.JavaPlugin;
+import org.gestern.gringotts.accountholder.AccountHolderFactory;
+import org.gestern.gringotts.accountholder.AccountHolderProvider;
 import org.gestern.gringotts.api.impl.VaultConnector;
+import org.gestern.gringotts.event.AccountListener;
+import org.gestern.gringotts.event.PlayerVaultListener;
+import org.gestern.gringotts.event.VaultCreator;
 import org.mcstats.MetricsLite;
-
-import static org.gestern.gringotts.Configuration.CONF;
-
 
 
 public class Gringotts extends JavaPlugin {
@@ -31,8 +35,12 @@ public class Gringotts extends JavaPlugin {
 
     /** Manages accounts. */
     public Accounting accounting;
-
-
+    
+    /** 
+     * The account holder factory is the place to go if you need an AccountHolder instance for an id.
+     */
+    public AccountHolderFactory accountHolderFactory = new AccountHolderFactory();
+    
     @Override
     public void onEnable() {
     	
@@ -47,19 +55,11 @@ public class Gringotts extends JavaPlugin {
 	        CONF.readConfig(savedConfig);
 	    	
 	    	gcommand = new Commands(this);
-	    	 
-	        CommandExecutor playerCommands = gcommand.new Money();
-	        CommandExecutor moneyAdminCommands = gcommand.new Moneyadmin();
-	        CommandExecutor adminCommands = gcommand.new GringottsCmd();
-	
-	        getCommand("balance").setExecutor(playerCommands);
-	        getCommand("money").setExecutor(playerCommands);
-	        getCommand("moneyadmin").setExecutor(moneyAdminCommands);
-	        getCommand("gringotts").setExecutor(adminCommands);
 	
 	        accounting = new Accounting();
-	
-	        getServer().getPluginManager().registerEvents(new AccountListener(this), this);
+	        
+	        registerCommands();
+	        registerEvents();
 	        registerEconomy();
 	        
 	        try {
@@ -103,6 +103,27 @@ public class Gringotts extends JavaPlugin {
         
         log.info("disabled");
     }
+	
+	private void registerCommands() {
+		CommandExecutor playerCommands = gcommand.new Money();
+        CommandExecutor moneyAdminCommands = gcommand.new Moneyadmin();
+        CommandExecutor adminCommands = gcommand.new GringottsCmd();
+        
+		getCommand("balance").setExecutor(playerCommands);
+        getCommand("money").setExecutor(playerCommands);
+        getCommand("moneyadmin").setExecutor(moneyAdminCommands);
+        getCommand("gringotts").setExecutor(adminCommands);
+	}
+	
+	private void registerEvents() {
+		PluginManager manager = getServer().getPluginManager();
+		
+		manager.registerEvents(new AccountListener(), this);
+		manager.registerEvents(new PlayerVaultListener(), this);
+		manager.registerEvents(new VaultCreator(), this);
+		
+		// listeners for other account types are loaded with dependencies
+	}
 
 
 	/**
@@ -116,6 +137,10 @@ public class Gringotts extends JavaPlugin {
 		} else {
 			log.info("Vault not found. Other plugins may not be able to access Gringotts accounts.");
 		}
+	}
+	
+	public void registerAccountHolderProvider(String type, AccountHolderProvider provider) {
+		accountHolderFactory.registerAccountHolderProvider(type, provider);
 	}
 
 }