+++ src/main/java/org/gestern/gringotts/accountholder/AccountHolderFactory.java
@@ -1,12 +1,10 @@
 package org.gestern.gringotts.accountholder;
 
-import static org.gestern.gringotts.dependency.Dependency.DEP;
-
-import java.util.logging.Logger;
+import java.util.LinkedHashMap;
+import java.util.Map;
 
 import org.bukkit.Bukkit;
 import org.bukkit.OfflinePlayer;
-import org.gestern.gringotts.Gringotts;
 
 /**
  * Manages creating various types of AccountHolder centrally.
@@ -16,34 +14,28 @@ import org.gestern.gringotts.Gringotts;
  */
 public class AccountHolderFactory {
 	
-	private final Logger log = Gringotts.G.getLogger();
+	private final Map<String, AccountHolderProvider> accountHolderProviders;
+	
+	public AccountHolderFactory() {
+		// linked hashmap maintains iteration order -> prefer player to be checked first
+		accountHolderProviders = new LinkedHashMap<String, AccountHolderProvider>();
+		accountHolderProviders.put("player", new PlayerAccountHolderProvider());
+		
+        // TODO support banks
+        // TODO support virtual accounts
+	}
 
 	/**
 	 * Get an account holder with automatically determined type, based on the owner's name.
 	 * @param owner
 	 * @return account holder for the given owner name, or null if none could be determined
 	 */
-    public AccountHolder get(String owner) {    	
-        OfflinePlayer player = Bukkit.getOfflinePlayer(owner);
-        // if this player has ever played on the server, they are a legit account holder
-        if (player.isOnline() || player.hasPlayedBefore()) {
-            return new PlayerAccountHolder(player);
-        }
-
-        if (DEP.factions.enabled()) {
-        	AccountHolder holder = DEP.factions.getAccountHolderByName(owner);
-        	if (holder != null) return holder;
-        }
+    public AccountHolder get(String owner) {
         
-        if (DEP.towny.enabled()) {
-    		AccountHolder holder = DEP.towny.getAccountHolderByAccountName(owner);
-    		if (holder!=null) return holder;
+        for (AccountHolderProvider provider : accountHolderProviders.values()) {
+        	AccountHolder accountHolder = provider.getAccountHolder(owner);
+        	if (accountHolder != null) return accountHolder;
         }
-        
-        // TODO support banks
-        // TODO support virtual accounts
-
-        log.fine("No account holder found for " + owner);
         
         return null;
     }
@@ -55,29 +47,32 @@ public class AccountHolderFactory {
      * @return account holder of given type with given owner name, or null if none could be determined or type is not supported.
      */
     public AccountHolder get(String type, String owner) {
-    	if (type.equals("player")) {
-            OfflinePlayer player = Bukkit.getOfflinePlayer(owner);
-            // if this player has ever played on the server, they are a legit account holder
-            if (player.isOnline() || player.getLastPlayed() > 0) {
-                return new PlayerAccountHolder(player);
-            }
-            else return null;
-    	}
+
+    	AccountHolderProvider provider = accountHolderProviders.get(type);
     	
-    	if (DEP.factions.enabled() && type.equals("faction")) {
-    		AccountHolder holder = DEP.factions.getAccountHolderById(owner); // try by id first, then by "faction-id" form name
-    		if (holder==null) holder = DEP.factions.getAccountHolderByName(owner); // maybe a bit hacky? whatever.
-    		if (holder!=null) return holder;
-    	}
+    	AccountHolder accountHolder = null;
+    	if (provider != null)
+    		accountHolder = provider.getAccountHolder(owner);
     	
-    	if (DEP.towny.enabled() && type.equals("towny")) {
-    		AccountHolder holder = DEP.towny.getAccountHolderByAccountName(owner);
-    		if (holder!=null) return holder;
-    	} 
-    	
-    	// no valid type
-    	return null;
+    	return accountHolder;
+    }
+
+	public void registerAccountHolderProvider(String type, AccountHolderProvider provider) {
+	    accountHolderProviders.put(type, provider);
     }
     
+	private static class PlayerAccountHolderProvider implements AccountHolderProvider {
+
+		@Override
+        public AccountHolder getAccountHolder(String id) {
+			OfflinePlayer player = Bukkit.getOfflinePlayer(id);
+	        // if this player has ever played on the server, they are a legit account holder
+	        if (player.isOnline() || player.hasPlayedBefore())
+	            return new PlayerAccountHolder(player);
+	        else 
+	        	return null;
+        }
+		
+	}
 
 }