+++ src/main/java/org/gestern/gringotts/dependency/FactionsHandler.java
@@ -1,7 +1,11 @@
 package org.gestern.gringotts.dependency;
 
+import org.bukkit.Bukkit;
 import org.bukkit.entity.Player;
+import org.gestern.gringotts.Gringotts;
+import org.gestern.gringotts.accountholder.AccountHolderProvider;
 import org.gestern.gringotts.accountholder.FactionAccountHolder;
+import org.gestern.gringotts.event.FactionsListener;
 
 import com.massivecraft.factions.FPlayer;
 import com.massivecraft.factions.FPlayers;
@@ -9,12 +13,15 @@ import com.massivecraft.factions.Faction;
 import com.massivecraft.factions.Factions;
 import com.massivecraft.factions.P;
 
-public class FactionsHandler implements DependencyHandler {
+public class FactionsHandler implements DependencyHandler, AccountHolderProvider {
 	
 	private final P plugin;
 	
 	public FactionsHandler(P plugin) {
 		this.plugin = plugin;
+		
+		Bukkit.getPluginManager().registerEvents(new FactionsListener(), Gringotts.G);
+		Gringotts.G.registerAccountHolderProvider("faction", this);
 	}
 
 	/**
@@ -27,34 +34,6 @@ public class FactionsHandler implements DependencyHandler {
         FPlayer fplayer = FPlayers.i.get(player);
         Faction playerFaction = fplayer.getFaction();
         return playerFaction != null? new FactionAccountHolder(playerFaction) : null;
-	}
-	
-	/**
-	 * Get a FactionAccountHolder based on the name of the account. 
-	 * Only names beginning with "faction-" will be considered, and the rest of the string 
-	 * can be either a faction id or a faction tag.
-	 * @param name Name of the account.
-	 * @return a FactionAccountHolder based on the name of the account, if a valid faction could be found. null otherwise.
-	 */
-	public FactionAccountHolder getAccountHolderByName(String name) {
-    	
-		// only a valid faction account name if owner starts with "faction-"
-		if ( ! name.startsWith("faction-")) return null;
-				
-		// not sure, but somehow this is sometimes id, sometimes tag??
-        String factionTag = name.substring(8);
-        Faction faction;
-        
-        // try id first
-        faction = Factions.i.get(factionTag);            
-        // and then tag
-        if (faction == null)
-        	faction = Factions.i.getByTag(factionTag);
-        
-        if (faction != null) 
-            return new FactionAccountHolder(faction);
-        
-        return null;
 	}
 	
 	/**
@@ -75,6 +54,34 @@ public class FactionsHandler implements DependencyHandler {
 	@Override
     public boolean exists() {
 	    return plugin != null;
+    }
+
+	/**
+	 * Get a FactionAccountHolder based on the name of the account. 
+	 * Only names beginning with "faction-" will be considered, and the rest of the string 
+	 * can be either a faction id or a faction tag.
+	 * @param name Name of the account.
+	 * @return a FactionAccountHolder based on the name of the account, if a valid faction could be found. null otherwise.
+	 */
+	@Override
+    public FactionAccountHolder getAccountHolder(String id) {
+		// only a valid faction account name if owner starts with "faction-"
+		if ( ! id.startsWith("faction-")) return null;
+				
+		// not sure, but somehow this is sometimes id, sometimes tag??
+        String factionTag = id.substring(8);
+        Faction faction;
+        
+        // try id first
+        faction = Factions.i.get(factionTag);            
+        // and then tag
+        if (faction == null)
+        	faction = Factions.i.getByTag(factionTag);
+        
+        if (faction != null) 
+            return new FactionAccountHolder(faction);
+        
+        return null;
     }
 
 }