+++ src/main/java/org/gestern/gringotts/event/PlayerVaultCreationEvent.java
@@ -1,0 +1,33 @@
+package org.gestern.gringotts.event;
+
+import org.bukkit.event.HandlerList;
+import org.bukkit.event.block.SignChangeEvent;
+
+/**
+ * Vault creation event triggered by a player.
+ * @author jast
+ *
+ */
+public class PlayerVaultCreationEvent extends VaultCreationEvent {
+
+	private final SignChangeEvent cause;
+	
+	public PlayerVaultCreationEvent(String type, SignChangeEvent cause) {
+	    super(type);
+	    this.cause = cause;
+    }
+	
+	/**
+	 * Get the player involved in creating the vault.
+	 * @return
+	 */
+	public SignChangeEvent getCause() {
+		return cause;
+	}
+	
+	public static HandlerList getHandlerList() {
+	    return VaultCreationEvent.handlers;
+	    // TODO ensure we can actually have superclass handle these safely
+	}
+
+}