+++ changelog.md
@@ -1,6 +1,12 @@
 Gringotts Changelog
 ===================
 
+v1.6.0
+------
+* can create a vault now by attaching the [vault] sign to a chest directly
+* for devs: added an event system for vault creation. Gringotts throws a `PlayerVaultCreationEvent` when a player attempts
+to create a vault in the world. Plugins can register an event handler and supply their information for Gringotts to create their own vault types.
+
 v1.5.0
 ------
 * add commands `/money withdraw` and `/money deposit` to withdraw money from vaults to inventory and vice versa.